web/codebases: propagate github connect button props
Created with Sturdy
